requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
newsapi-python==0.2.6
psycopg2==2.8.6
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
pytz==2021.1
requests==2.25.1
rope==0.18.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
